% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerArchive.R
\name{buildPlayerArchive}
\alias{buildPlayerArchive}
\title{Title Builds the usage stats database}
\usage{
buildPlayerArchive(pathServer, pathArch, ...)
}
\arguments{
\item{pathServer}{The url to the main hlstats page}

\item{pathArch}{The path where the output sql database should be}

\item{...}{Passthrough for customization 'getHlTopPlayers' or the rankings}
}
\value{
A logical if the sql was disconnected, side effect of writing the
  database
}
\description{
Driver that scrapes the session data and then builds the models for tryhard
}

