% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlayerTendencies.R
\name{calcPlayerTendencies}
\alias{calcPlayerTendencies}
\title{Title Calculate the models for the usage statistics}
\usage{
calcPlayerTendencies(sess)
}
\arguments{
\item{sess}{data.frame of the events for a single player on the server gotten
by using getSessionTimes}
}
\value{
a list with two models for tryhard and lifetime
}
\description{
Title Calculate the models for the usage statistics
}

