% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getServer.R
\name{getHlServerPage}
\alias{getHlServerPage}
\title{Driver to scrape the hlstats main pages}
\usage{
getHlServerPage(pathServer, boxName = "\\\\S", type = c("header",
  "table")[1], shOnlyNodes = FALSE)
}
\arguments{
\item{pathServer}{The url to the main hlstats page}

\item{boxName}{The name of the area that you want to capture}

\item{type}{regex of the box that you want captured}

\item{shOnlyNodes}{Return the xml nodes or the tables}
}
\value{
Either the xml nodes or a string matrix of the table
}
\description{
Driver to scrape the hlstats main pages
}

