% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getServer.R
\name{getSessionTimes}
\alias{getSessionTimes}
\title{Parse the data on each players per login stats for usage statistics}
\usage{
getSessionTimes(pathServer, playerId, verbose = c("all", "min",
  "onlyDate")[1])
}
\arguments{
\item{pathServer}{The url to the main hlstats page}

\item{playerId}{The player's id}

\item{verbose}{Which columns to return as follows:
\describe{
  \item{all}{All the columns}
  \item{min}{Only returns the date, time and skill change}
  \item{onlyDate}{Only returns the date}
}}
}
\description{
Parse the data on each players per login stats for usage statistics
}

