% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerArchive.R
\name{queryPlayerArchive}
\alias{queryPlayerArchive}
\title{Title Get the usage stats for a given table from the sql database}
\usage{
queryPlayerArchive(currentTab, pathArch)
}
\arguments{
\item{currentTab}{Data.frame of the server page from getServerPage}

\item{pathArch}{The path where the output sql database should be}
}
\value{
list with predictions for tryhard and lifetime for all the players in
  the server
}
\description{
Title Get the usage stats for a given table from the sql database
}

