% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhlstats.R
\docType{package}
\name{rhlstats-package}
\alias{rhlstats}
\alias{rhlstats-package}
\title{Scrapes and analyzes HLStatsX video game servers}
\description{
HLStatsX servers meticulous track usage rates in a number of
categories. To see the health of a server, these stats should be easily
downloaded and analyzed. This package sets up an archive for this data and does
some further statistics on how each player uses the server. This culminates
with a measure of how alive the server is at the moment and over a certain time
period with "lifetime" and "tryhard".
}
\details{
...
}

