% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getServer.R
\name{getHlTopPlayers}
\alias{getHlTopPlayers}
\title{Grab the top player lists for the first server}
\usage{
getHlTopPlayers(pathServer, numPlayers = 50, rankingOrder = "skill",
  rankingTime = c("Total", "Week", "Month")[1], shJustIds = FALSE)
}
\arguments{
\item{pathServer}{The url to the main hlstats page}

\item{numPlayers}{The number of players to download(grabs in 50 player pages)}

\item{rankingOrder}{What variable should it be ranked by}

\item{rankingTime}{What time period should be considered for the ranking}

\item{shJustIds}{Give only the player ids of the players}
}
\value{
A dataframe of the final table or a vector of the player Ids
}
\description{
Grab the top player lists for the first server
}

